admin access for CRUD order

+++ src/main/java/com/dell/webservice/controller/OrderController.java
@@ -18,12 +18,16 @@ import org.springframework.web.bind.annotation.RestController;
 import com.dell.webservice.entity.Order;
 import com.dell.webservice.entity.Product;
 import com.dell.webservice.repository.OrderService;
+import com.dell.webservice.repository.UserService;
 
 @RestController
 public class OrderController {
 	//get orders
 	@Autowired
 	OrderService orderService;
+	
+	@Autowired
+	UserService userService; 
 	
 			@GetMapping("/orders")
 			public ResponseEntity<?> getOrder(
@@ -40,14 +44,19 @@ public class OrderController {
 			
 		//create a product
 			@PostMapping("/orders")
-			public ResponseEntity<?> addOrder(@RequestBody(required = false) Order orderObj){
+			public ResponseEntity<?> addOrder(@RequestBody(required = false) Order orderObj, @RequestParam(required = false) String userName){
 				if(orderObj==null)
 				{
 					return new ResponseEntity<String>("Request body can not be empty", new HttpHeaders(), HttpStatus.BAD_REQUEST);
 				}
 				try {
-					 this.orderService.addOrder(orderObj);
-					 return new ResponseEntity<List<Product>>( new HttpHeaders(), HttpStatus.CREATED);
+					boolean role = userService.checkAdmin(userName);
+					 if(role == true) {
+						 this.orderService.addOrder(orderObj);
+						 return new ResponseEntity<List<Product>>( new HttpHeaders(), HttpStatus.CREATED);
+					 }else {
+						 return new ResponseEntity<String>("Unauthorized Request",new HttpHeaders(), HttpStatus.UNAUTHORIZED);
+					 } 
 				}catch(Exception ex) {
 					return new ResponseEntity<String>("Unable to add orders", new HttpHeaders(),HttpStatus.INTERNAL_SERVER_ERROR);
 				}
@@ -55,7 +64,7 @@ public class OrderController {
 			
 			//update a order
 					@PutMapping("/orders/{id}")
-					public ResponseEntity<?> updateOrder(@PathVariable("id") int id, @RequestBody(required = false) Order orderObj){
+					public ResponseEntity<?> updateOrder(@PathVariable("id") int id, @RequestBody(required = false) Order orderObj, @RequestParam(required = false) String userName){
 						if(orderObj==null)
 						{
 							return new ResponseEntity<String>("Request body can not be empty", new HttpHeaders(), HttpStatus.BAD_REQUEST);
@@ -63,31 +72,46 @@ public class OrderController {
 						if(id != orderObj.getId()) {   
 							return new ResponseEntity<String>("Id in requested path and requested body do not match", new HttpHeaders(), HttpStatus.BAD_REQUEST);
 						}
-						try {
-							Order fetchOrder = this.orderService.getOrderById(id);
-							if(fetchOrder == null)
-							{
-								return new ResponseEntity<String>("Order does not exist with  id"+id, new HttpHeaders(), HttpStatus.NOT_FOUND);
-							}
-							 this.orderService.updateOrder(orderObj);
-							 return new ResponseEntity<String>("Order is added successfully!", new HttpHeaders(), HttpStatus.NO_CONTENT);
-						}catch(Exception ex) 
+						try{
+							boolean role = userService.checkAdmin(userName);
+						 if(role == true) {
+							 Order fetchOrder = this.orderService.getOrderById(id);
+								if(fetchOrder == null)
+								{
+									return new ResponseEntity<String>("Order does not exist with  id"+id, new HttpHeaders(), HttpStatus.NOT_FOUND);
+								}
+								 this.orderService.updateOrder(orderObj);
+								 return new ResponseEntity<String>("Order is added successfully!", new HttpHeaders(), HttpStatus.NO_CONTENT);
+						
+						 }else {
+							 return new ResponseEntity<String>("Unauthorized Request",new HttpHeaders(), HttpStatus.UNAUTHORIZED);
+						 }
+						
+								}catch(Exception ex) 
 						{
 							return new ResponseEntity<String>("Unable to add order", new HttpHeaders(),HttpStatus.INTERNAL_SERVER_ERROR);
 						}
 					}
 					
-			//delete order
+					//delete order
 					@DeleteMapping("/orders/{id}")
-					public ResponseEntity<?> deleteOrder(@PathVariable("id") int id){
+					public ResponseEntity<?> deleteOrder(@PathVariable("id") int id, @RequestParam(required = false) String userName){
 						try {
-							Order fetchOrder = this.orderService.getOrderById(id);
-							if(fetchOrder == null) {
-								return new ResponseEntity<String>("Order does not exist with id"+id,new HttpHeaders(), HttpStatus.NOT_FOUND);
-							}else {
-								this.orderService.deleteOrder(id);
-								return new ResponseEntity<Product>(new HttpHeaders(), HttpStatus.NO_CONTENT);
+							boolean role = this.userService.checkAdmin(userName);
+							if(role == true)
+							{
+								Order fetchOrder = this.orderService.getOrderById(id);
+								if(fetchOrder == null) {
+									return new ResponseEntity<String>("Order does not exist with id"+id,new HttpHeaders(),
+											 HttpStatus.NOT_FOUND);
+								}else{
+									this.orderService.deleteOrder(id);
+									return new ResponseEntity<Product>(new HttpHeaders(), HttpStatus.NO_CONTENT);
+								}
+							}
+							else {
 								
+								return new ResponseEntity<String>("Unauthorized Request", new HttpHeaders(), HttpStatus.UNAUTHORIZED);
 							}
 						}catch(Exception ex) {
 							return new ResponseEntity<String>("Unable to delete order", new HttpHeaders(), HttpStatus.INTERNAL_SERVER_ERROR);
